# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a pending order and initialize payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Pending order with Paystack authorization URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          description: Payment provider is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order status and purchased books
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{orderId}/retry:
    post:
      operationId: retryOrderPayment
      tags: [orders]
      summary: Retry payment for a pending order
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: New payment session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Order is already paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Payment provider is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}/downloads:
    get:
      operationId: listOrderDownloads
      tags: [orders]
      summary: List downloads unlocked for a paid order
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: Download links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Download"
        "403":
          description: Order is not paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /payments/paystack/webhook:
    post:
      operationId: paystackWebhook
      tags: [orders]
      summary: Confirm Paystack payments server-side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaystackWebhook"
      responses:
        "200":
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "401":
          description: Invalid Paystack signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Sign in to the private admin panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Check the current admin session
      responses:
        "200":
          description: Current admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get admin dashboard totals and recent orders
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/books:
    get:
      operationId: listAdminBooks
      tags: [admin]
      summary: List all books for administration
      responses:
        "200":
          description: Books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createBook
      tags: [admin]
      summary: Add an ebook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Created book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/books/{bookId}:
    patch:
      operationId: updateBook
      tags: [admin]
      summary: Edit an ebook
      parameters:
        - $ref: "#/components/parameters/BookIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteBook
      tags: [admin]
      summary: Remove an ebook from the catalogue
      parameters:
        - $ref: "#/components/parameters/BookIdPath"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List customer orders
      parameters:
        - $ref: "#/components/parameters/StatusQuery"
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /admin/orders/{orderId}:
    get:
      operationId: getAdminOrder
      tags: [admin]
      summary: Get an order for administration
      parameters:
        - $ref: "#/components/parameters/OrderIdPath"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a private ebook or cover upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Presigned upload details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  parameters:
    BookIdPath:
      name: bookId
      in: path
      required: true
      schema:
        type: string
    OrderIdPath:
      name: orderId
      in: path
      required: true
      schema:
        type: string
...
    StatusQuery:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [pending, paid, failed]
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Admin authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Book:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        author: { type: string }
        price: { type: number }
        currency: { type: string }
        category: { type: string }
        description: { type: string }
        format: { type: string, enum: [PDF, EPUB] }
        coverUrl: { type: ["string", "null"] }
        fileName: { type: string }
        featured: { type: boolean }
        publishedAt: { type: string }
        createdAt: { type: string }
      required: [id, slug, title, author, price, currency, category, description, format, coverUrl, fileName, featured, publishedAt, createdAt]
    BookInput:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        author: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        category: { type: string, minLength: 1 }
        description: { type: string }
        format: { type: string, enum: [PDF, EPUB] }
        coverObjectPath: { type: ["string", "null"] }
        fileObjectPath: { type: string, minLength: 1 }
        fileName: { type: string, minLength: 1 }
        featured: { type: boolean }
        publishedAt: { type: string }
      required: [title, slug, author, price, currency, category, description, format, coverObjectPath, fileObjectPath, fileName, featured, publishedAt]
    BookUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        author: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        category: { type: string, minLength: 1 }
        description: { type: string }
        format: { type: string, enum: [PDF, EPUB] }
        coverObjectPath: { type: ["string", "null"] }
        fileObjectPath: { type: string, minLength: 1 }
        fileName: { type: string, minLength: 1 }
        featured: { type: boolean }
        publishedAt: { type: string }
...
    OrderInput:
      type: object
      properties:
        email: { type: string, minLength: 3 }
        country: { type: string, minLength: 2, maxLength: 2 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        bookIds: { type: array, minItems: 1, items: { type: string } }
      required: [email, country, currency, bookIds]
    CheckoutSession:
      type: object
      properties:
        orderId: { type: string }
        reference: { type: string }
        authorizationUrl: { type: string }
        accessCode: { type: string }
      required: [orderId, reference, authorizationUrl, accessCode]
    Order:
      type: object
      properties:
        id: { type: string }
        reference: { type: string }
        email: { type: string }
        country: { type: string }
        currency: { type: string }
        subtotal: { type: number }
        status: { type: string, enum: [pending, paid, failed] }
        paymentStatus: { type: string }
        deliveryEmailSent: { type: boolean }
        downloaded: { type: boolean }
        createdAt: { type: string }
        paidAt: { type: ["string", "null"] }
        items: { type: array, items: { $ref: "#/components/schemas/OrderItem" } }
      required: [id, reference, email, country, currency, subtotal, status, paymentStatus, deliveryEmailSent, downloaded, createdAt, paidAt, items]
    OrderItem:
      type: object
      properties:
        bookId: { type: string }
        title: { type: string }
        author: { type: string }
        price: { type: number }
        format: { type: string }
        downloadUrl: { type: ["string", "null"] }
        downloaded: { type: boolean }
      required: [bookId, title, author, price, format, downloadUrl, downloaded]
    Download:
      type: object
      properties:
        bookId: { type: string }
        title: { type: string }
        format: { type: string }
        url: { type: string }
      required: [bookId, title, format, url]
    PaystackWebhook:
      type: object
      properties:
        event: { type: string }
        data:
          type: object
          properties:
            reference: { type: string }
            status: { type: string }
          required: [reference]
      required: [event, data]
    WebhookResponse:
      type: object
      properties:
        received: { type: boolean }
      required: [received]
    AdminLoginInput:
      type: object
      properties:
        email: { type: string, minLength: 3 }
        password: { type: string, minLength: 1 }
      required: [email, password]
    AdminSession:
      type: object
      properties:
        authenticated: { type: boolean }
        email: { type: ["string", "null"] }
      required: [authenticated, email]
    AdminDashboard:
      type: object
      properties:
        totalOrders: { type: number }
        pendingOrders: { type: number }
        totalBooks: { type: number }
        recentOrders: { type: array, items: { $ref: "#/components/schemas/Order" } }
      required: [totalOrders, pendingOrders, totalBooks, recentOrders]
    UploadInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        size: { type: number, minimum: 1 }
        contentType: { type: string, minLength: 1 }
      required: [name, size, contentType]
    UploadResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [uploadURL, objectPath]