# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookInputFormat } from './bookInputFormat';

export interface BookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  author: string;
  /** @minimum 0 */
  price: number;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency: string;
  /** @minLength 1 */
  category: string;
  description: string;
  format: BookInputFormat;
  /** @nullable */
  coverObjectPath: string | null;
  /** @minLength 1 */
  fileObjectPath: string;
  /** @minLength 1 */
  fileName: string;
  featured: boolean;
  publishedAt: string;
}