# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  reference: string;
  email: string;
  country: string;
  currency: string;
  subtotal: number;
  status: OrderStatus;
  paymentStatus: string;
  deliveryEmailSent: boolean;
  downloaded: boolean;
  createdAt: string;
  /** @nullable */
  paidAt: string | null;
  items: OrderItem[];
}